# Eliminate arithmetic instructions based on properties