@jserv 2025-03-26
The results generated by `hyperfine --warmup 3 'out/shecc src/main.c'` looks weird. Can you utilize tools such as [uftrace](https://github.com/namhyung/uftrace) to figure out the slowdown?