@ChAoSUnItY 2025-03-26
The result given by `uftrace` is contrary to the result generated by `hyperfine`, the optimized one is faster then the original one.
![image](https://github.com/user-attachments/assets/357ab560-59a9-4dcb-9913-03ebce7360e9)