@ChAoSUnItY 2025-03-26
No, and I don't have any armv7 / riscv machine to run tests. I am currently using qemu to bootstrap and benchmark locally, which may not be accurate.